+++ Packages/idv.jlchntoz.vvmw/Runtime/VVMW/Designer/ActiveRegionConfig.cs
@@ -24,6 +24,7 @@ namespace JLChnToZ.VRC.VVMW {
         internal Core core;
         [SerializeField, LocalizedLabel] internal Bounds bounds;
         [SerializeField, LocalizedLabel] internal bool staticRegion = true;
+        [SerializeField, LocalizedLabel] internal bool useWorldSpaceBounds;
         [NonSerialized] Core lastCore;
 
         Core IVizVidCompoonent.Core => core;